added error checking and other changes based on reviewer comments
