+++ sycl/test-e2e/ESIMD/PerformanceTests/BitonicSortK.cpp
@@ -5,11 +5,11 @@
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 //
 //===----------------------------------------------------------------------===//
-// REQUIRES: gpu-intel-dg2
+// REQUIRES: gpu-intel-dg2 && level_zero
 
 // RUN: mkdir -p %t.dir && %{build} -o %t.dir/exec.out
 // RUN: env IGC_DumpToCustomDir=%t.dir IGC_ShaderDumpEnable=1 %{run} %t.dir/exec.out
 // RUN: python3 %S/instruction_count.py %t.dir 3452 VC_asmfc04983569d0d4c9__ZTSZZN11BitonicSort5SolveEPjS0_jENKUlRN4sycl3_V17handlerEE0_clES4_E5Merge.asm
-// RUN: rm -rf %t.dir
+// RUN: echo "Baseline from driver version 1.3.29138"
 
 #include "../BitonicSortK.cpp"
